refactor: remove `hc` prefix from module names
